% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymp_gee.R
\name{asymp_gee}
\alias{asymp_gee}
\title{Asymptotic Power Computation for Generalised Estimating Equations (GEE)}
\usage{
asymp_gee(
  n = 1000,
  m = 4,
  p = 0.3,
  a = 0.9,
  tau1 = 0.08,
  tau2 = 0.07,
  delta = 3,
  d = 0.005
)
}
\arguments{
\item{n}{numeric.}

\item{m}{numeric.}

\item{p}{numeric.}

\item{a}{numeric.}

\item{tau1}{numeric.}

\item{tau2}{numeric.}

\item{delta}{numeric.}

\item{d}{numeric.}
}
\value{
numeric.
}
\description{
Asymptotic Power Computation for Generalised Estimating Equations (GEE)
}
