% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymp_lmm.R
\name{asymp_lmm}
\alias{asymp_lmm}
\title{Asymptotic Power Computation for Linear Mixed Effects Model (LMM)}
\usage{
asymp_lmm(
  n = 1000,
  p = 0.3,
  m = 4,
  delta = 3,
  sig2 = 0.08,
  sig2_b1 = 8e-04,
  d = 0.005
)
}
\arguments{
\item{n}{numeric.}

\item{p}{numeric.}

\item{m}{numeric.}

\item{delta}{numeric.}

\item{sig2}{numeric.}

\item{sig2_b1}{numeric.}

\item{d}{numeric.}
}
\value{
numeric.
}
\description{
Asymptotic Power Computation for Linear Mixed Effects Model (LMM)
}
