% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palgwas-package.R
\docType{package}
\name{palgwas-package}
\alias{palgwas}
\alias{palgwas-package}
\title{palgwas: Power Analysis for Longitudinal Genome-Wide Association Studies}
\description{
Shiny App to explore statistical power for longitudinal 
    Genome-Wide Association Studies using various parameters and two approach, namely,
    Generalised Estimating Equations and Linear Mixed effect Model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcanouil/palgwas}
  \item Report bugs at \url{https://github.com/mcanouil/palgwas/issues}
}

}
\author{
\strong{Maintainer}: Mickaël Canouil \email{mickael.canouil@cnrs.fr} (\href{https://orcid.org/0000-0002-3396-4549}{ORCID})

Authors:
\itemize{
  \item Ghislain Rocheleau
}

}
\keyword{internal}
